feat: Material Symbolsアイコン実装とセクション名変更
主な変更点:
- SVGアイコンをMaterial Symbolsに統一実装
- アイコン背景を円形から四角形（border-radius: 4px）に変更
- Material Symbolsフォントをlayout.txaのheadタグ内で直接読み込み
- InterviewRecordsコンポーネントをSummaryにリネーム（セクション名も「要約」に変更）
- 全セクションアイコンをMaterial Symbols化: person, description, campaign, contact_page, groups, share
- 連絡先アイコン（email, phone）もMaterial Symbols対応
- MaterialIcons.tsxファイル追加（一時的なフォント読み込み用）
