+++ src/app/globals.css
@@ -60,8 +60,28 @@ body {
   }
   
   .icon-contact {
+    font-family: 'Material Symbols Outlined', sans-serif;
+    font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 20;
     color: var(--brand-primary);
-    @apply w-4 h-4;
+    font-size: 12px;
+    width: 20px;
+    height: 20px;
+    background: var(--brand-white);
+    border-radius: 50%;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    font-style: normal;
+    line-height: 1;
+    text-transform: none;
+    letter-spacing: normal;
+    word-wrap: normal;
+    white-space: nowrap;
+    direction: ltr;
+    -webkit-font-smoothing: antialiased;
+    text-rendering: optimizeLegibility;
+    -moz-osx-font-smoothing: grayscale;
+    font-feature-settings: 'liga';
   }
   
   .card-content-padding {
@@ -216,5 +236,30 @@ body {
   
   .activity-badge {
     @apply text-xs px-2 py-0.5 rounded-full border;
+  }
+  
+  .material-icon {
+    font-family: 'Material Symbols Outlined', sans-serif;
+    font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 20;
+    color: var(--brand-primary);
+    font-size: 16px;
+    width: 32px;
+    height: 32px;
+    background: var(--brand-lighter-blue);
+    border-radius: 4px;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    font-style: normal;
+    line-height: 1;
+    text-transform: none;
+    letter-spacing: normal;
+    word-wrap: normal;
+    white-space: nowrap;
+    direction: ltr;
+    -webkit-font-smoothing: antialiased;
+    text-rendering: optimizeLegibility;
+    -moz-osx-font-smoothing: grayscale;
+    font-feature-settings: 'liga';
   }
 }
