+++ src/app/layout.tsx
@@ -25,6 +25,12 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ja">
+      <head>
+        <link 
+          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" 
+          rel="stylesheet" 
+        />
+      </head>
       <body
         className={`${inter.variable} ${notoSansJP.variable} antialiased`}
       >
