+++ src/components/profile/ActivityInfo.tsx
@@ -13,11 +13,7 @@ export function ActivityInfo({ activities }: ActivityInfoProps) {
       <CardHeader className="card-header-padding">
         <div className="flex items-center justify-between">
           <CardTitle className="card-title-base">
-            <div className="icon-base">
-              <svg viewBox="0 0 24 24" fill="currentColor" className="w-full h-full">
-                <path d="M6.503 20.752c0 1.794-1.456 3.248-3.251 3.248-1.796 0-3.252-1.454-3.252-3.248 0-1.794 1.456-3.248 3.252-3.248 1.795.001 3.251 1.454 3.251 3.248zm-6.503-12.572v4.811c6.05.062 10.96 4.966 11.022 11.009h4.818c-.062-8.71-7.118-15.758-15.84-15.82zm0-3.368c10.58.046 19.152 8.594 19.183 19.188h4.817c-.03-13.231-10.755-23.954-24-24v4.812z" />
-              </svg>
-            </div>
+            <span className="material-icon">campaign</span>
             活動情報
           </CardTitle>
           <Button
