+++ src/components/profile/CareerInfo.tsx
@@ -12,11 +12,7 @@ export function CareerInfo({ currentInfo, pastBusinessCards }: CareerInfoProps) 
     <Card className="profile-card">
       <CardHeader className="card-header-padding">
         <CardTitle className="card-title-base">
-          <div className="icon-base">
-            <svg viewBox="0 0 24 24" fill="currentColor" className="w-full h-full">
-              <path d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z" />
-            </svg>
-          </div>
+          <span className="material-icon">person</span>
           経歴情報
         </CardTitle>
       </CardHeader>
